re_cl/requirements.txt
# RE_CL — Plataforma de Detección de Inmuebles Subvalorados
# Python 3.11+

# Base de datos
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
alembic==1.13.1

# ETL / Data
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1

# GIS
geopandas==0.14.4
shapely==2.0.4
pyproj==3.6.1
folium==0.16.0

# ML / Scoring
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.0

# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

# Dashboard
streamlit==1.35.0
plotly==5.22.0

# Testing
pytest==8.2.0
pytest-cov==5.0.0
httpx==0.27.0   # Para test de FastAPI

# Utilidades
tqdm==4.66.4
loguru==0.7.2

# V4 — Backtesting OLS benchmark
statsmodels>=0.14.0

# V4 — OSM Overpass API requests
requests>=2.31.0

# V5 — Financial simulator
numpy_financial>=1.0.0

# V2 — Orchestration
prefect==2.19.9

# V2 — Scraping
playwright==1.45.0          # después de instalar: playwright install chromium

# V2 — Frontend dev server (opcional, el frontend usa npm)
# npm install en re_cl/frontend/ para React + Deck.gl

# V6 — Excel export
openpyxl>=3.1.0

# V6.6 — JWT auth
python-jose[cryptography]>=3.3.0
passlib>=1.7.4                  # uses pbkdf2_sha256 (bcrypt extra has Py3.13 issues)
